+++ libopendmi/src/stream.c
@@ -6,17 +6,17 @@
 //
 #include <memory.h>
 
+#include <opendmi/entity.h>
 #include <opendmi/stream.h>
 
-bool dmi_stream_initialize(dmi_stream_t *stream, const void *data, size_t length)
+bool dmi_stream_initialize(dmi_stream_t *stream, const dmi_entity_t *entity)
 {
-    if ((stream == nullptr) or (data == nullptr))
+    if ((stream == nullptr) or (entity == nullptr))
         return false;
 
-    stream->data      = data;
-    stream->length    = length;
+    stream->entity    = entity;
     stream->position  = 0;
-    stream->remaining = length;
+    stream->remaining = entity->body_length;
 
     return true;
 }
@@ -25,11 +25,11 @@ bool dmi_stream_seek(dmi_stream_t *stream, size_t position)
 {
     if (stream == nullptr)
         return false;
-    if (position >= stream->length)
+    if (position >= stream->entity->body_length)
         return false;
 
     stream->position  = position;
-    stream->remaining = stream->length - position;
+    stream->remaining = stream->entity->body_length - position;
 
     return true;
 }
@@ -52,7 +52,7 @@ bool dmi_stream_peek_data(dmi_stream_t *stream, void *ptr, size_t length)
     if (stream->remaining < length)
         return false;
 
-    memcpy(ptr, stream->data + stream->position, length);
+    memcpy(ptr, stream->entity->data + stream->position, length);
 
     return true;
 }
@@ -60,5 +60,5 @@ bool dmi_stream_peek_data(dmi_stream_t *stream, void *ptr, size_t length)
 void dmi_stream_reset(dmi_stream_t *stream)
 {
     stream->position  = 0;
-    stream->remaining = stream->length;
+    stream->remaining = stream->entity->body_length;
 }